requirements.txt
# Supabase (primary backend)
supabase>=2.28.0

# Core ML
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0       # YOLOv8

# Geospatial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
pyproj>=3.5.0

# Data & AWS
boto3>=1.26.0
awscli>=1.27.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0

# Roboflow
roboflow>=1.2.0

# Active Learning / Uncertainty
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
Pillow>=10.0.0

# Inference Server
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic>=2.0.0          # Used directly in inference_server.py (transitive via fastapi, but declared explicitly)
python-multipart>=0.0.9

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
